[Fixed Bean Issue]
