--- PluggableAuth/src/main/PluggableAuth/Adapters/inbound/EmailLogin/EmailLogin.java
@@ -1,12 +1,0 @@
-package com.Group25.PluggableAuth.Adapters.inbound.EmailLogin;
-
-public class EmailLogin {
-    private String email;
-
-    public String getEmail(){
-        return email;
-    }
-    public void setEmail(String email){
-        this.email = email;
-    }
-}
