--- PluggableAuth/src/main/PluggableAuth/PluggableAuthApplication.java
@@ -1,13 +1,0 @@
-package com.Group25.PluggableAuth;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PluggableAuthApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PluggableAuthApplication.class, args);
-	}
-
-}
