--- PluggableAuth/src/main/PluggableAuth/Port/EmailPort.java
@@ -1,4 +1,0 @@
-package com.Group25.PluggableAuth.Port;
-public interface EmailPort{
-    int sendMail(String email, String messageText);
-}
