+++ PluggableAuth/src/main/java/com/Group25/PluggableAuth/Adapters/outbound/OutboundAdapterConfig.java
@@ -1,5 +1,18 @@
 package com.Group25.PluggableAuth.Adapters.outbound;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.Group25.PluggableAuth.Adapters.outbound.SendMail.SendMailDemo;
+import com.Group25.PluggableAuth.Port.EmailPort;
+
+@Configuration
 public class OutboundAdapterConfig {
+
+@Bean
+    public EmailPort emailClient() {
+        return new SendMailDemo();
+    }
     
 }
