+++ PluggableAuth/src/main/java/com/Group25/PluggableAuth/Adapters/outbound/SendMail/SendMailDemo.java
@@ -68,7 +68,7 @@ public class SendMailDemo implements EmailPort {
             }
         
         }
-    
+        scanner.close();
     }
 
         catch(FileNotFoundException e){ //Handles file not found error
